refactor(session): extract FileLoaderMixin to _session_files.py
Behavior-preserving move of open_file/open_stack/load_example loaders,
the nav-shape prompt round-trip, save methods, and the module-level file
helpers (SUPPORTED_EXTS, _path_ext, _is_supported_dataset_path,
_dataset_size_bytes, …). Helpers re-imported into session.py so
'from spyde.backend.session import _path_ext' etc. still resolve.
session.py drops 2000 -> 445 lines.
